+++ pyaud/_utils.py
@@ -9,12 +9,8 @@ from __future__ import annotations
 import typing as _t
 from pathlib import Path as _Path
 
-import setuptools as _setuptools
 from gitspy import Git as _Git
 from object_colors import Color as _Color
-
-from . import config as _config
-from . import exceptions as _exceptions
 
 colors = _Color()
 git = _Git()
@@ -47,23 +43,7 @@ def get_packages() -> _t.List[str]:
         found.
     :return: List of Python packages.
     """
-    packages = list(
-        set(
-            i.split(".", maxsplit=1)[0]
-            for i in _setuptools.find_packages(
-                # in response to an update to `setuptools` stubs:
-                # - error: Argument "where" has incompatible type
-                #   "Path"; expected "str"
-                where=str(_Path.cwd()),
-                exclude=_config.toml["packages"]["exclude"],
-            )
-        )
-    )
-    if not packages:
-        raise _exceptions.PythonPackageNotFoundError("no packages found")
-
-    packages.sort()
-    return packages
+    return [_Path.cwd().name]
 
 
 def package() -> str:
@@ -75,29 +55,7 @@ def package() -> str:
     """
     # at least one package will be returned or an error would have been
     # raised
-    packages = get_packages()
-
-    # if there is only one package then that is the default
-    if len(packages) == 1:
-        return packages.pop()
-
-    # if there are multiple packages found then look for a configured
-    # package name that matches one of the project's packages
-    package_name = _config.toml["packages"].get("name")
-    if package_name in packages:
-        return package_name
-
-    # if there are multiple packages found, and none of the above two
-    # apply, then the package with the same name as the project root (if
-    # it exists) is the default
-    repo = _Path.cwd().name
-    if repo in packages:
-        return repo
-
-    # if none of the above criteria is met then raise
-    raise _exceptions.PythonPackageNotFoundError(
-        "cannot determine primary package"
-    )
+    return get_packages()[0]
 
 
 def get_commit_hash() -> _t.Optional[str]: